// sdk/contoso/arm-contoso/src/models/models.ts
export interface Resource {
  /** Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName} */
  readonly id?: string;
  /** The name of the resource */
  readonly name?: string;
  /** The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts" */
  readonly type?: string;
  /** Azure Resource Manager metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
}
// ...
export interface TrackedResource extends Resource {
  /** Resource tags. */
  tags?: Record<string, string>;
  /** The geo-location where the resource lives */
  location: string;
}

export function trackedResourceSerializer(
  item: TrackedResource,
): TrackedResourceRest {
  return {
    tags: !item.tags ? item.tags : (serializeRecord(item.tags as any) as any),
    location: item["location"],
  };
}

/** Employee resource */
export interface Employee extends TrackedResource {
  /** The resource-specific properties for this resource. */
  properties?: EmployeeProperties;
}

export function employeeSerializer(item: Employee): EmployeeRest {
  return {
    tags: !item.tags ? item.tags : (serializeRecord(item.tags as any) as any),
    location: item["location"],
    properties: !item.properties
      ? item.properties
      : employeePropertiesSerializer(item.properties),
  };
}

/** Employee properties */
export interface EmployeeProperties {
  /** Age of employee */
  age?: number;
  /** City of employee */
  city?: string;
  /** Profile of employee */
  profile?: Uint8Array;
  /** The status of the last operation. */
  readonly provisioningState?: ProvisioningState;
}
// ...
export interface _EmployeeListResult {
  /** The Employee items on this page */
  value: Employee[];
  /** The link to the next page of items */
  nextLink?: string;
}
// ...
export type ProvisioningState =
  | ResourceProvisioningState
  | "Provisioning"
  | "Updating"
  | "Deleting"
  | "Accepted"
  | string;